+++ frontend/appflowy_flutter/pubspec.yaml
@@ -162,7 +162,7 @@ dependency_overrides:
   appflowy_editor:
     git:
       url: https://github.com/AppFlowy-IO/appflowy-editor.git
-      ref: "82f3baf"
+      ref: "739cd59"
 
   logger: ^2.0.0
